+++ src/react/components/editabletable/XUIEditableTableCellTextInput.js
@@ -24,6 +24,7 @@ const XUIEditableTableCellTextInput = ({
   const textInputComponentRef = React.useRef();
   const [isFocused, setIsFocused] = React.useState();
   const [timeoutHandle, setTimeoutHandle] = React.useState(null);
+  const [isFocusingViaClick, setIsFocusingViaClick] = React.useState(false);
 
   /**
    * Records the focus state onBlur, before calling any user-supplied handlers.
@@ -34,7 +35,7 @@ const XUIEditableTableCellTextInput = ({
 
     clearTimeout(timeoutHandle);
 
-    onBlur && onBlur(event);
+    onBlur?.(event);
   };
 
   /**
@@ -45,11 +46,20 @@ const XUIEditableTableCellTextInput = ({
   const composedOnFocus = event => {
     setIsFocused(true);
 
-    const { target } = event;
-    // Timeout must be used to avoid a synchronisation issue with Safari
-    setTimeoutHandle(setTimeout(() => target.select(), 0));
+    onFocus?.(event);
+  };
 
-    onFocus && onFocus(event);
+  const onMouseDown = () => {
+    setIsFocusingViaClick(!isFocused);
+  };
+
+  const onClick = event => {
+    if (isFocusingViaClick && getSelection().toString().length === 0) {
+      const { target } = event;
+      // Timeout must be used to avoid a synchronisation issue with Safari
+      setTimeoutHandle(setTimeout(() => target.select(), 0));
+    }
+    setIsFocusingViaClick(false);
   };
 
   /**
@@ -84,7 +94,9 @@ const XUIEditableTableCellTextInput = ({
         isLabelHidden
         isValueReverseAligned={isValueReverseAligned || inlineAlignment === 'end'}
         onBlur={composedOnBlur}
+        onClick={onClick}
         onFocus={composedOnFocus}
+        onMouseDown={onMouseDown}
         ref={textInputComponentRef}
       />
     </XUIEditableTableCellControl>