+++ src/react/components/range/XUIRange.js
@@ -41,7 +41,7 @@ const XUIRange = props => {
     labelRequiredError(
       XUIRange.name,
       ['includes a label with text', 'labelId provided'],
-      [labelRef.current?.innerText && !isLabelHidden, typeof label?.[0] === 'string', labelId],
+      [labelRef.current?.textContent && !isLabelHidden, typeof label?.[0] === 'string', labelId],
     );
   }, [isLabelHidden, label, labelId, labelRef]);